[feat] 패킷 최적화, SToCSessionManager::Broadcast()의 중대한 보틀넥 수정, TestClient 구체적인 실행 정보 출력하도록 기능 제작, 미들맨 SToL과 SToC 스레드 개수 비율 조정

+++ IocpLib/NetCodes/Session.h
@@ -39,8 +39,8 @@ public:
 	SOCKET GetSocket() { return _socket; }
 	bool IsConnected() { return _connected; }
 	SharedPtr<Session> GetSessionRef() { return static_pointer_cast<Session>(shared_from_this()); }
-	const uint64 GetSessionId() { return _sessionId; }
-	bool SetSessionId(uint64 id) { _sessionId = id; return true; }
+	const uint16 GetSessionId() { return _sessionId; }
+	bool SetSessionId(uint16 id) { _sessionId = id; return true; }
 
 private:
 	virtual HANDLE GetHandle() override;
@@ -83,7 +83,7 @@ private:
 	DisconnectEvent _disconnectEvent;
 	RecvEvent _recvEvent;
 	SendEvent _sendEvent;
-	uint64 _sessionId = 0; // 0: 초기화 안됨
+	uint16 _sessionId = 0; // 0: 초기화 안됨
 };
 
 class PacketSession : public Session
