+++ MainServer/SToCSessionManager.h
@@ -12,11 +12,11 @@ public:
 	void Add(SharedPtr<SToCSession> session);
 	void Remove(SharedPtr<SToCSession> session);
 	void Broadcast(SharedPtr<SendBuffer> sendBuffer);
-	void Send(uint64 sessionId, SharedPtr<SendBuffer> sendBuffer);
+	void Send(uint16 sessionId, SharedPtr<SendBuffer> sendBuffer);
 
 private:
 	USE_LOCK;
-	HashMap<uint64, SharedPtr<SToCSession>> _sessions;
+	HashMap<uint16, SharedPtr<SToCSession>> _sessions;
 };
 
 extern SToCSessionManager GSToCSessionManager;
